fix: AuthenticationClaimsExtractor ClassCastException 수정
