+++ backend/build.gradle.kts
@@ -12,7 +12,7 @@ plugins {
 }
 
 group = "com"
-version = "2.2.0-SNAPSHOT"
+version = "2.2.1-SNAPSHOT"
 
 java {
     sourceCompatibility = JavaVersion.VERSION_17
