+++ backend/src/main/kotlin/com/festago/auth/config/LoginConfig.kt
@@ -3,10 +3,10 @@ package com.festago.auth.config
 import com.festago.auth.annotation.Authorization
 import com.festago.auth.domain.Role
 import com.festago.auth.domain.authentication.AuthenticateContext
+import com.festago.auth.domain.token.http.CompositeHttpRequestTokenExtractor
+import com.festago.auth.domain.token.http.CookieHttpRequestTokenExtractor
+import com.festago.auth.domain.token.http.HeaderHttpRequestTokenExtractor
 import com.festago.auth.domain.token.jwt.AuthenticationTokenExtractor
-import com.festago.auth.infrastructure.token.http.CompositeHttpRequestTokenExtractor
-import com.festago.auth.infrastructure.token.http.CookieHttpRequestTokenExtractor
-import com.festago.auth.infrastructure.token.http.HeaderHttpRequestTokenExtractor
 import com.festago.auth.web.argument.AdminAuthenticationArgumentResolver
 import com.festago.auth.web.argument.MemberAuthenticationArgumentResolver
 import com.festago.auth.web.argument.RoleArgumentResolver
