+++ backend/src/main/kotlin/com/festago/auth/domain/token/jwt/extractor/AdminAuthenticationTokenExtractor.kt
@@ -1,19 +1,19 @@
-package com.festago.auth.infrastructure.token.jwt
+package com.festago.auth.domain.token.jwt.extractor
 
 import com.festago.auth.domain.authentication.Authentication
 import com.festago.auth.domain.token.jwt.AuthenticationTokenExtractor
 import com.festago.auth.domain.token.jwt.JwtTokenParser
-import com.festago.auth.infrastructure.token.jwt.claims.MemberAuthenticationClaimsExtractor
+import com.festago.auth.domain.token.jwt.extractor.claims.AdminAuthenticationClaimsExtractor
 import org.springframework.stereotype.Component
 
 @Component
-internal class MemberAuthenticationTokenExtractor(
+internal class AdminAuthenticationTokenExtractor(
     private val jwtTokenParser: JwtTokenParser,
-    private val memberAuthenticationClaimsExtractor: MemberAuthenticationClaimsExtractor,
+    private val adminAuthenticationClaimsExtractor: AdminAuthenticationClaimsExtractor,
 ) : AuthenticationTokenExtractor {
 
     override fun extract(token: String): Authentication {
         val claims = jwtTokenParser.getClaims(token)
-        return memberAuthenticationClaimsExtractor.extract(claims)
+        return adminAuthenticationClaimsExtractor.extract(claims)
     }
 }
