+++ backend/src/main/kotlin/com/festago/auth/domain/token/jwt/extractor/MemberAuthenticationTokenExtractor.kt
@@ -1,19 +1,19 @@
-package com.festago.auth.infrastructure.token.jwt
+package com.festago.auth.domain.token.jwt.extractor
 
 import com.festago.auth.domain.authentication.Authentication
 import com.festago.auth.domain.token.jwt.AuthenticationTokenExtractor
 import com.festago.auth.domain.token.jwt.JwtTokenParser
-import com.festago.auth.infrastructure.token.jwt.claims.AdminAuthenticationClaimsExtractor
+import com.festago.auth.domain.token.jwt.extractor.claims.MemberAuthenticationClaimsExtractor
 import org.springframework.stereotype.Component
 
 @Component
-internal class AdminAuthenticationTokenExtractor(
+internal class MemberAuthenticationTokenExtractor(
     private val jwtTokenParser: JwtTokenParser,
-    private val adminAuthenticationClaimsExtractor: AdminAuthenticationClaimsExtractor,
+    private val memberAuthenticationClaimsExtractor: MemberAuthenticationClaimsExtractor,
 ) : AuthenticationTokenExtractor {
 
     override fun extract(token: String): Authentication {
         val claims = jwtTokenParser.getClaims(token)
-        return adminAuthenticationClaimsExtractor.extract(claims)
+        return memberAuthenticationClaimsExtractor.extract(claims)
     }
 }
