+++ backend/src/test/kotlin/com/festago/auth/domain/token/jwt/extractor/CompositeAuthenticationTokenExtractorTest.kt
@@ -1,10 +1,10 @@
-package com.festago.auth.infrastructure.token.jwt
+package com.festago.auth.domain.token.jwt.extractor
 
 import com.festago.auth.domain.Role
 import com.festago.auth.domain.authentication.AdminAuthentication
 import com.festago.auth.domain.authentication.AnonymousAuthentication
 import com.festago.auth.domain.token.jwt.JwtTokenParser
-import com.festago.auth.domain.token.jwt.claims.AuthenticationClaimsExtractor
+import com.festago.auth.domain.token.jwt.extractor.claims.AuthenticationClaimsExtractor
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
